React Bootstrap - Section 4
Imported React Bootstrap and used the Card, Row and Col display features to list the products on the home page. Created a component for Product to display image and title on each card. Also created a component for Rating to display the rating and number of reviews on each card. Use a link from Font Awesome to show the ratings and stars and half-stars.