Product Screen - Section 5
Implemented more details in the Product Screen component and set it to 3 columns. Added a new API on the server to return the products details screen based on it's slug. Installed react-helmet-async to update the browser tab (title) info. Created a function in the utilities file to handle the display of error messages when the user types in an invalid web page. Created components for MessageBox and LoadingBox that is used across different components.